Move ProductsController management actions to Foodouken directory and enable authorization.
